+++ imports/api/create-pdf/proposal/observations/index.js
@@ -4,7 +4,7 @@ export default function conditions (props) {
   if (!props.observations.external.trim()) return null;
 
   return [
-    {text: 'Condições da Negociação', alignment: 'center', style: 'h2'},
+    {text: 'Descrição da Proposta', alignment: 'center', style: 'h2'},
     {text: props.observations.external, style: 'p'}
   ]
 }