+++ imports/api/create-pdf/proposal/signature/index.js
@@ -4,7 +4,7 @@ export default function signature (data) {
   return {text: [
           `${moment().format("DD/MM/YYYY")}\n`,
           'Atenciosamente,\n',
-          'Jurgen Junior\n',
+          data.createdByFullName + "\n",
           'Fone: (11) 5532-0790 / (11) 55335614'
       ], alignment: 'right', style: 'p'}
 }